make-a-thon demo version
Gyro_controller class was not used for the demo because it bricks the Arduino board.  See readme for resetting the board if a similar issue occurred. 
buttons were not supported in the program.
The rotary encoder will need one extra interrupt pin to make it work properly. Now it only works in one direction.
More documentation of the current programs will be updated to the readme.